requirements.txt
asgiref==3.8.1
black==24.10.0
bleach==6.1.0
click==8.1.7
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-summernote==0.8.20.0
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
six==1.16.0
sqlparse==0.5.1
webencodings==0.5.1
